Rollback Implement of changeable column names.

+++ src/SimpleUser/UserManager.php
@@ -42,25 +42,6 @@ class UserManager implements UserProviderInterface
     /** @var string */
     protected $userCustomFieldsTableName = 'user_custom_fields';
 
-    /** @var array */
-    protected $userColumns = array(
-        'id' => 'id',
-        'email' => 'email',
-        'password' => 'password',
-        'salt' => 'salt',
-        'roles' => 'roles',
-        'name' => 'name',
-        'time_created' => 'time_created',
-        'username' => 'username',
-        'isEnabled' => 'isEnabled',
-        'confirmationToken' => 'confirmationToken',
-        'timePasswordResetRequested' => 'timePasswordResetRequested',
-        //Custom Fields
-        'user_id' => 'user_id',
-        'attribute' => 'attribute',
-        'value' => 'value',
-    );
-
     /**
      * Constructor.
      *
@@ -88,7 +69,7 @@ class UserManager implements UserProviderInterface
     public function loadUserByUsername($username)
     {
         if (strpos($username, '@') !== false) {
-            $user = $this->findOneBy(array($this->getUserColumns('email') => $username));
+            $user = $this->findOneBy(array('email' => $username));
             if (!$user) {
                 throw new UsernameNotFoundException(sprintf('Email "%s" does not exist.', $username));
             }
@@ -96,7 +77,7 @@ class UserManager implements UserProviderInterface
             return $user;
         }
 
-        $user = $this->findOneBy(array($this->getUserColumns('username') => $username));
+        $user = $this->findOneBy(array('username' => $username));
         if (!$user) {
             throw new UsernameNotFoundException(sprintf('Username "%s" does not exist.', $username));
         }
@@ -149,12 +130,7 @@ class UserManager implements UserProviderInterface
     {
         // Test for new columns added in v2.0.
         // If they're missing, throw an exception and explain that migration is needed.
-        foreach (array(
-                    $this->getUserColumns('username'),
-                    $this->getUserColumns('isEnabled'),
-                    $this->getUserColumns('confirmationToken'),
-                    $this->getUserColumns('timePasswordResetRequested')
-                ) as $col) {
+        foreach (array('username', 'isEnabled', 'confirmationToken', 'timePasswordResetRequested') as $col) {
             if (!array_key_exists($col, $data)) {
                 throw new \RuntimeException('Internal error: database schema appears out of date. See https://github.com/jasongrimes/silex-simpleuser/blob/master/sql/MIGRATION.md');
             }
@@ -354,7 +330,7 @@ class UserManager implements UserProviderInterface
      */
     public function getUser($id)
     {
-        return $this->findOneBy(array($this->getUserColumns('id') => $id));
+        return $this->findOneBy(array('id' => $id));
     }
 
     /**
@@ -387,9 +363,8 @@ class UserManager implements UserProviderInterface
     public function findBy(array $criteria = array(), array $options = array())
     {
         // Check the identity map first.
-        if (array_key_exists($this->getUserColumns('id'), $criteria) 
-            && array_key_exists($criteria[$this->getUserColumns('id')], $this->identityMap)) {
-            return array($this->identityMap[$criteria[$this->getUserColumns('id')]]);
+        if (array_key_exists('id', $criteria) && array_key_exists($criteria['id'], $this->identityMap)) {
+            return array($this->identityMap[$criteria['id']]);
         }
 
         list ($common_sql, $params) = $this->createCommonFindSql($criteria);
@@ -409,10 +384,10 @@ class UserManager implements UserProviderInterface
 
         $users = array();
         foreach ($data as $userData) {
-            if (array_key_exists($userData[$this->getUserColumns('id')], $this->identityMap)) {
-                $user = $this->identityMap[$userData[$this->getUserColumns('id')]];
+            if (array_key_exists($userData['id'], $this->identityMap)) {
+                $user = $this->identityMap[$userData['id']];
             } else {
-                $userData['customFields'] = $this->getUserCustomFields($userData[$this->getUserColumns('id')]);
+                $userData['customFields'] = $this->getUserCustomFields($userData['id']);
                 $user = $this->hydrateUser($userData);
                 $this->identityMap[$user->getId()] = $user;
             }
@@ -432,7 +407,7 @@ class UserManager implements UserProviderInterface
 
         $rows = $this->conn->fetchAll('SELECT * FROM ' . $this->conn->quoteIdentifier($this->userCustomFieldsTableName). ' WHERE user_id = ?', array($userId));
         foreach ($rows as $row) {
-            $customFields[$row[$this->getUserColumns('attribute')]] = $row[$this->getUserColumns('value')];
+            $customFields[$row['attribute']] = $row['value'];
         }
 
         return $customFields;
@@ -456,9 +431,9 @@ class UserManager implements UserProviderInterface
                 $i++;
                 $alias = 'custom' . $i;
                 $sql .= 'JOIN ' . $this->conn->quoteIdentifier($this->userCustomFieldsTableName). ' ' . $alias . ' ';
-                $sql .= 'ON ' . $this->conn->quoteIdentifier($this->userTableName). '.' . $this->getUserColumns('id') . ' = ' . $alias . '.'. $this->getUserColumns('user_id').' ';
-                $sql .= 'AND ' . $alias . '.'.$this->getUserColumns('attribute').' = :attribute' . $i . ' ';
-                $sql .= 'AND ' . $alias . '.'.$this->getUserColumns('value').' = :value' . $i . ' ';
+                $sql .= 'ON ' . $this->conn->quoteIdentifier($this->userTableName). '.id = ' . $alias . '.user_id ';
+                $sql .= 'AND ' . $alias . '.attribute = :attribute' . $i . ' ';
+                $sql .= 'AND ' . $alias . '.value = :value' . $i . ' ';
                 $params['attribute' . $i] = $attribute;
                 $params['value' . $i] = $value;
             }
@@ -503,9 +478,7 @@ class UserManager implements UserProviderInterface
         $this->dispatcher->dispatch(UserEvents::BEFORE_INSERT, new UserEvent($user));
 
         $sql = 'INSERT INTO ' . $this->conn->quoteIdentifier($this->userTableName) . '
-            ('.$this->getUserColumns('email').', '.$this->getUserColumns('password').', '.$this->getUserColumns('salt').', '.$this->getUserColumns('name').
-                ', '.$this->getUserColumns('roles').', '.$this->getUserColumns('time_created').', '.$this->getUserColumns('username').', '.$this->getUserColumns('isEnabled').
-                ', '.$this->getUserColumns('confirmationToken').', '.$this->getUserColumns('timePasswordResetRequested').')
+            (email, password, salt, name, roles, time_created, username, isEnabled, confirmationToken, timePasswordResetRequested)
             VALUES (:email, :password, :salt, :name, :roles, :timeCreated, :username, :isEnabled, :confirmationToken, :timePasswordResetRequested) ';
 
         $params = array(
@@ -542,17 +515,17 @@ class UserManager implements UserProviderInterface
         $this->dispatcher->dispatch(UserEvents::BEFORE_UPDATE, new UserEvent($user));
 
         $sql = 'UPDATE ' . $this->conn->quoteIdentifier($this->userTableName). '
-            SET '.$this->getUserColumns('email').' = :email
-            , '.$this->getUserColumns('password').' = :password
-            , '.$this->getUserColumns('salt').' = :salt
-            , '.$this->getUserColumns('name').' = :name
-            , '.$this->getUserColumns('roles').' = :roles
-            , '.$this->getUserColumns('time_created').' = :timeCreated
-            , '.$this->getUserColumns('username').' = :username
-            , '.$this->getUserColumns('isEnabled').' = :isEnabled
-            , '.$this->getUserColumns('confirmationToken').' = :confirmationToken
-            , '.$this->getUserColumns('timePasswordResetRequested').' = :timePasswordResetRequested
-            WHERE '.$this->getUserColumns('id').' = :id';
+            SET email = :email
+            , password = :password
+            , salt = :salt
+            , name = :name
+            , roles = :roles
+            , time_created = :timeCreated
+            , username = :username
+            , isEnabled = :isEnabled
+            , confirmationToken = :confirmationToken
+            , timePasswordResetRequested = :timePasswordResetRequested
+            WHERE id = :id';
 
         $params = array(
             'email' => $user->getEmail(),
@@ -580,12 +553,10 @@ class UserManager implements UserProviderInterface
      */
     protected function saveUserCustomFields(User $user)
     {
-        $this->conn->executeUpdate('DELETE FROM ' . $this->conn->quoteIdentifier($this->userCustomFieldsTableName). ' 
-            WHERE '.$this->getUserColumns('user_id').' = ?', array($user->getId()));
+        $this->conn->executeUpdate('DELETE FROM ' . $this->conn->quoteIdentifier($this->userCustomFieldsTableName). ' WHERE user_id = ?', array($user->getId()));
 
         foreach ($user->getCustomFields() as $attribute => $value) {
-            $this->conn->executeUpdate('INSERT INTO ' . $this->conn->quoteIdentifier($this->userCustomFieldsTableName). 
-                    ' ('.$this->getUserColumns('user_id').', '.$this->getUserColumns('attribute').', '.$this->getUserColumns('value').') VALUES (?, ?, ?) ',
+            $this->conn->executeUpdate('INSERT INTO ' . $this->conn->quoteIdentifier($this->userCustomFieldsTableName). ' (user_id, attribute, value) VALUES (?, ?, ?) ',
                 array($user->getId(), $attribute, $value));
         }
     }
@@ -601,8 +572,8 @@ class UserManager implements UserProviderInterface
 
         $this->clearIdentityMap($user);
 
-        $this->conn->executeUpdate('DELETE FROM ' . $this->conn->quoteIdentifier($this->userTableName). ' WHERE '.$this->getUserColumns('id').' = ?', array($user->getId()));
-        $this->conn->executeUpdate('DELETE FROM ' . $this->conn->quoteIdentifier($this->userCustomFieldsTableName). ' WHERE '.$this->getUserColumns('user_id').' = ?', array($user->getId()));
+        $this->conn->executeUpdate('DELETE FROM ' . $this->conn->quoteIdentifier($this->userTableName). ' WHERE id = ?', array($user->getId()));
+        $this->conn->executeUpdate('DELETE FROM ' . $this->conn->quoteIdentifier($this->userCustomFieldsTableName). ' WHERE user_id = ?', array($user->getId()));
 
         $this->dispatcher->dispatch(UserEvents::AFTER_DELETE, new UserEvent($user));
     }
@@ -621,7 +592,7 @@ class UserManager implements UserProviderInterface
         $errors = $user->validate();
 
         // Ensure email address is unique.
-        $duplicates = $this->findBy(array($this->getUserColumns('email') => $user->getEmail()));
+        $duplicates = $this->findBy(array('email' => $user->getEmail()));
         if (!empty($duplicates)) {
             foreach ($duplicates as $dup) {
                 if ($user->getId() && $dup->getId() == $user->getId()) {
@@ -632,7 +603,7 @@ class UserManager implements UserProviderInterface
         }
 
         // Ensure username is unique.
-        $duplicates = $this->findBy(array($this->getUserColumns('username') => $user->getRealUsername()));
+        $duplicates = $this->findBy(array('username' => $user->getRealUsername()));
         if (!empty($duplicates)) {
             foreach ($duplicates as $dup) {
                 if ($user->getId() && $dup->getId() == $user->getId()) {
@@ -705,22 +676,6 @@ class UserManager implements UserProviderInterface
         return $this->userTableName;
     }
 
-    public function setUserColumns(array $userColumns){
-        $conn = $this->conn;
-        //Escape the column names
-
-        $escapedUserColumns = array_map(function($column) use ($conn){
-            return $conn->quoteIdentifier($column,\PDO::PARAM_STR);
-        }, $userColumns);
-
-        //Merge the existing column names
-        $this->userColumns = array_merge($this->userColumns, $escapedUserColumns);
-    }
-
-    public function getUserColumns($column = ""){
-        if ($column == "") return $this->userColumns;
-        else return $this->userColumns[$column];
-    }
 
     public function setUserCustomFieldsTableName($userCustomFieldsTableName)
     {
